+++ PassWard/CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+### 1.1.0
+
+  * Fixed for the `v0.220.3` patch.
+  * Bumped up `<LangVersion>` to C# 12.
+  * Code clean-up and refactoring.
+  * Added new config-option `HoverText.userListSeparator`.
+
 ### 1.0.0
 
   * Initial release.